requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
psycopg[binary]>=3.1
